% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.data.R
\docType{data}
\name{iso_codes}
\alias{iso_codes}
\title{Countries and codes}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 286 rows and 2 columns.
}
\usage{
iso_codes
}
\description{
This dataset maps countries to codes.

For information about the ISO standard for country codes see
\url{https://www.iso.org/iso-3166-country-codes.html}.
}
\section{Definitions}{


\itemize{
\item \code{country} (character): Country name., * \code{country_iso} (character): Corresponding ISO code.
}

}

\examples{
head(iso_codes)
}
\seealso{
Other Datasets: 
\code{\link{gics_classification}},
\code{\link{increasing_or_decreasing}},
\code{\link{isic_classification}},
\code{\link{market_share}},
\code{\link{nace_classification}},
\code{\link{naics_classification}},
\code{\link{palette_colours}},
\code{\link{psic_classification}},
\code{\link{region_isos}},
\code{\link{scenario_colours}},
\code{\link{sda}},
\code{\link{sector_classifications}},
\code{\link{sector_colours}},
\code{\link{sic_classification}},
\code{\link{technology_colours}}
}
\concept{Datasets}
\keyword{datasets}
