% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot.R
\docType{data}
\name{market_share}
\alias{market_share}
\title{An example of a \code{market_share}-like dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 802 rows and 10 columns.
}
\usage{
market_share
}
\description{
Dataset imitating the output of \code{\link[r2dii.analysis:target_market_share]{r2dii.analysis::target_market_share()}}.
}
\examples{
market_share
}
\seealso{
Other datasets: 
\code{\link{gics_classification}},
\code{\link{increasing_or_decreasing}},
\code{\link{isic_classification}},
\code{\link{iso_codes}},
\code{\link{nace_classification}},
\code{\link{naics_classification}},
\code{\link{palette_colours}},
\code{\link{psic_classification}},
\code{\link{region_isos}},
\code{\link{scenario_colours}},
\code{\link{sda}},
\code{\link{sector_classifications}},
\code{\link{sector_colours}},
\code{\link{sic_classification}},
\code{\link{technology_colours}}
}
\concept{datasets}
\keyword{datasets}
