% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{pacta_loanbook_deps}
\alias{pacta_loanbook_deps}
\title{List all pacta.loanbook dependencies}
\usage{
pacta_loanbook_deps(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
pacta.loanbook packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
List all pacta.loanbook dependencies
}
