% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{pacta_loanbook_logo}
\alias{pacta_loanbook_logo}
\title{The \code{{pacta.loanbook}} logo, using ASCII or Unicode characters}
\usage{
pacta_loanbook_logo(unicode = cli::is_utf8_output())
}
\arguments{
\item{unicode}{Whether to use Unicode symbols. Default is \code{TRUE}
on UTF-8 platforms.}
}
\description{
Use \code{\link[cli:ansi_strip]{cli::ansi_strip()}} to get rid of the colors.
}
\examples{
pacta_loanbook_logo()
}
\seealso{
Other utility functions: 
\code{\link{pacta_loanbook_conflicts}()},
\code{\link{pacta_loanbook_deps}()},
\code{\link{pacta_loanbook_packages}()},
\code{\link{pacta_loanbook_sitrep}()},
\code{\link{pacta_loanbook_update}()}
}
\concept{utility functions}
