% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pacta_loanbook_packages}
\alias{pacta_loanbook_packages}
\title{List all packages in pacta.loanbook}
\usage{
pacta_loanbook_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include pacta.loanbook in the list?}
}
\description{
List all packages in pacta.loanbook
}
\examples{
pacta_loanbook_packages()
}
