% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{pacta_loanbook_update}
\alias{pacta_loanbook_update}
\title{Update pacta.loanbook packages}
\usage{
pacta_loanbook_update(recursive = FALSE, repos = getOption("repos"))
}
\arguments{
\item{recursive}{If \code{TRUE}, will also list all dependencies of
pacta.loanbook packages.}

\item{repos}{The repositories to use to check for updates.
Defaults to \code{getOption("repos")}.}
}
\description{
This will check to see if all pacta.loanbook packages (and optionally, their
dependencies) are up-to-date, and will install after an interactive
confirmation.
}
\examples{
\dontrun{
pacta_loanbook_update()
}
}
\seealso{
Other Utility functions: 
\code{\link{pacta_loanbook_conflicts}()},
\code{\link{pacta_loanbook_deps}()},
\code{\link{pacta_loanbook_logo}()},
\code{\link{pacta_loanbook_packages}()},
\code{\link{pacta_loanbook_sitrep}()}
}
\concept{Utility functions}
