% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot.R
\docType{data}
\name{palette_colours}
\alias{palette_colours}
\title{Colour datasets}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 9 rows and 2 columns.
}
\usage{
palette_colours
}
\description{
All datasets have at least two columns:
\itemize{
\item \code{label}: Text label of the colour.
\item \code{hex}: Hex code of the colour.
}
}
\details{
In \code{scenario_colours}, colours are ordered from red to green to be used in
trajectory charts.
}
\examples{
palette_colours

scenario_colours

sector_colours

technology_colours
}
\seealso{
Other datasets: 
\code{\link{gics_classification}},
\code{\link{increasing_or_decreasing}},
\code{\link{isic_classification}},
\code{\link{iso_codes}},
\code{\link{market_share}},
\code{\link{nace_classification}},
\code{\link{naics_classification}},
\code{\link{psic_classification}},
\code{\link{region_isos}},
\code{\link{scenario_colours}},
\code{\link{sda}},
\code{\link{sector_classifications}},
\code{\link{sector_colours}},
\code{\link{sic_classification}},
\code{\link{technology_colours}}
}
\concept{datasets}
\keyword{datasets}
