% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.match.R
\name{prioritize_level}
\alias{prioritize_level}
\title{Arrange unique \code{level} values in default order of \code{priority}}
\usage{
prioritize_level(data)
}
\arguments{
\item{data}{A data frame, commonly the output of \code{\link[r2dii.match:match_name]{match_name()}}.}
}
\value{
A character vector of the default level priority per loan.
}
\description{
Arrange unique \code{level} values in default order of \code{priority}
}
\examples{
matched <- tibble::tibble(
  level = c(
    "intermediate_parent_1",
    "direct_loantaker",
    "direct_loantaker",
    "direct_loantaker",
    "ultimate_parent",
    "intermediate_parent_2"
  )
)
prioritize_level(matched)
}
\seealso{
Other matching functions: 
\code{\link{calculate_match_success_rate}()},
\code{\link{crucial_lbk}()},
\code{\link{match_name}()},
\code{\link{prioritize}()}
}
\concept{matching functions}
