% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.analysis.R, R/r2dii.match.R,
%   R/r2dii.plot.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{target_market_share}
\alias{target_sda}
\alias{crucial_lbk}
\alias{match_name}
\alias{prioritize}
\alias{prioritize_level}
\alias{plot_emission_intensity}
\alias{plot_techmix}
\alias{plot_trajectory}
\alias{prep_emission_intensity}
\alias{prep_techmix}
\alias{prep_trajectory}
\alias{qplot_emission_intensity}
\alias{qplot_techmix}
\alias{qplot_trajectory}
\alias{recode_metric_techmix}
\alias{recode_metric_trajectory}
\alias{scale_colour_r2dii}
\alias{scale_colour_r2dii_sector}
\alias{scale_colour_r2dii_tech}
\alias{scale_fill_r2dii}
\alias{scale_fill_r2dii_sector}
\alias{spell_out_technology}
\alias{theme_2dii}
\alias{to_title}
\title{Objects exported from other packages}
\concept{Analysis functions}
\concept{Matching functions}
\concept{Plotting functions}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{r2dii.analysis}{\code{\link[r2dii.analysis]{target_market_share}}, \code{\link[r2dii.analysis]{target_sda}}}

  \item{r2dii.match}{\code{\link[r2dii.match]{crucial_lbk}}, \code{\link[r2dii.match]{match_name}}, \code{\link[r2dii.match]{prioritize}}, \code{\link[r2dii.match]{prioritize_level}}}

  \item{r2dii.plot}{\code{\link[r2dii.plot]{plot_emission_intensity}}, \code{\link[r2dii.plot]{plot_techmix}}, \code{\link[r2dii.plot]{plot_trajectory}}, \code{\link[r2dii.plot]{prep_emission_intensity}}, \code{\link[r2dii.plot]{prep_techmix}}, \code{\link[r2dii.plot]{prep_trajectory}}, \code{\link[r2dii.plot]{qplot_emission_intensity}}, \code{\link[r2dii.plot]{qplot_techmix}}, \code{\link[r2dii.plot]{qplot_trajectory}}, \code{\link[r2dii.plot:to_title]{recode_metric_techmix}}, \code{\link[r2dii.plot:to_title]{recode_metric_trajectory}}, \code{\link[r2dii.plot]{scale_colour_r2dii}}, \code{\link[r2dii.plot]{scale_colour_r2dii_sector}}, \code{\link[r2dii.plot]{scale_colour_r2dii_tech}}, \code{\link[r2dii.plot]{scale_colour_r2dii_tech}}, \code{\link[r2dii.plot:scale_colour_r2dii]{scale_fill_r2dii}}, \code{\link[r2dii.plot:scale_colour_r2dii_sector]{scale_fill_r2dii_sector}}, \code{\link[r2dii.plot:to_title]{spell_out_technology}}, \code{\link[r2dii.plot]{theme_2dii}}, \code{\link[r2dii.plot]{to_title}}}
}}

