% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot.R
\docType{data}
\name{sda_demo}
\alias{sda_demo}
\title{An example of an \code{sda_demo}-like dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 110 rows and 6 columns.
}
\source{
\url{https://github.com/RMI-PACTA/r2dii.plot/issues/55}.
}
\usage{
sda_demo
}
\description{
Dataset imitating the output of \code{\link[r2dii.analysis:target_sda]{r2dii.analysis::target_sda()}}.
}
\examples{
sda_demo
}
\seealso{
Other demo data: 
\code{\link{abcd_demo}},
\code{\link{co2_intensity_scenario_demo}},
\code{\link{loanbook_demo}},
\code{\link{market_share_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos_demo}},
\code{\link{scenario_demo_2020}}
}
\concept{demo data}
\keyword{datasets}
