% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r2dii.plot.R
\docType{data}
\name{market_share}
\alias{market_share}
\title{An example of a \code{market_share}-like dataset}
\format{
An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 802 rows and 10 columns.
}
\usage{
market_share
}
\description{
Dataset imitating the output of \code{\link[r2dii.analysis:target_market_share]{r2dii.analysis::target_market_share()}}.
}
\examples{
market_share
}
\seealso{
Other demo data: 
\code{\link{abcd_demo}},
\code{\link{co2_intensity_scenario_demo}},
\code{\link{loanbook_demo}},
\code{\link{overwrite_demo}},
\code{\link{region_isos_demo}},
\code{\link{scenario_demo_2020}},
\code{\link{sda}}
}
\concept{demo data}
\keyword{datasets}
