% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{pacta_loanbook_packages}
\alias{pacta_loanbook_packages}
\title{List all packages in \code{{pacta.loanbook}}}
\usage{
pacta_loanbook_packages(include_self = TRUE)
}
\arguments{
\item{include_self}{Include \code{{pacta.loanbook}} in the list?}
}
\value{
a character vector containing the name of packages imported by
\code{{pacta.loanbook}}.
}
\description{
List all packages in \code{{pacta.loanbook}}
}
\examples{
pacta_loanbook_packages()
}
\seealso{
Other utility functions: 
\code{\link{pacta_loanbook_conflicts}()},
\code{\link{pacta_loanbook_deps}()},
\code{\link{pacta_loanbook_logo}()},
\code{\link{pacta_loanbook_sitrep}()},
\code{\link{pacta_loanbook_update}()}
}
\concept{utility functions}
